+++ .github/workflows/promptflow-tracing-test.yml
@@ -44,15 +44,20 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
+      - uses: snok/install-poetry@v1
       - uses: actions/download-artifact@v4
         with:
           name: promptflow-tracing
+          path: ${{ env.WORKING_DIRECTORY }}
       - name: install promptflow-tracing from wheel
         # wildcard expansion (*) does not work in Windows, so leverage python to find and install
-        run: python -m pip install $(python -c "import glob; print(glob.glob('promptflow_tracing-*.whl')[0])")
-      - uses: snok/install-poetry@v1
+        run: poetry run pip install $(python -c "import glob; print(glob.glob('promptflow_tracing-*.whl')[0])")
+        working-directory: ${{ env.WORKING_DIRECTORY }}
       - name: install test dependency group
         run: poetry install --only test
+        working-directory: ${{ env.WORKING_DIRECTORY }}
+      - name: generate end-to-end test config from secret
+        run: echo '${{ secrets.PF_TRACING_E2E_TEST_CONFIG }}' >> connections.json
         working-directory: ${{ env.WORKING_DIRECTORY }}
       - name: test
         run: poetry run pytest
