+++ src/promptflow-tracing/promptflow/tracing/__init__.py
@@ -2,6 +2,8 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
+__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+
 from ._start_trace import start_trace
 from ._trace import trace
 from ._version import __version__
