+++ src/promptflow-tracing/promptflow/tracing/_tracer.py
@@ -1,5 +1,5 @@
-import json
 import inspect
+import json
 import logging
 import uuid
 from collections.abc import Iterator
