+++ src/promptflow-tracing/promptflow/tracing/_version.py
@@ -5,3 +5,5 @@
 import importlib.metadata
 
 __version__ = importlib.metadata.version("promptflow-tracing")
+
+VERSION: str = __version__
