+++ src/promptflow-tracing/promptflow/tracing/contracts/trace.py
