+++ src/promptflow-tracing/pyproject.toml
@@ -36,9 +36,9 @@ packages = [
 
 [tool.poetry.dependencies]
 python = "<4.0,>=3.8"
-openai = "*"
-tiktoken = ">=0.4.0"
-opentelemetry-sdk = ">=1.22.0,<2.0.0"
+openai = "*"  # API injector for OpenAI traces
+opentelemetry-sdk = ">=1.22.0,<2.0.0"  # Open Telemetry dependency
+tiktoken = ">=0.4.0"  # get OpenAI token count for streaming scenario
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "*"
@@ -46,7 +46,9 @@ import-linter = "*"
 
 [tool.poetry.group.test.dependencies]
 pytest = "*"
+pytest-asyncio = "*"
 pytest-cov = "*"
+pytest-mock = "*"
 pytest-xdist = "*"
 
 [tool.importlinter]
@@ -66,6 +68,7 @@ build-backend = "poetry.core.masonry.api"
 [tool.pytest.ini_options]
 markers = [
   "unittest",
+  "e2etest",
 ]
 # junit - analyse and publish test results (https://github.com/EnricoMi/publish-unit-test-result-action)
 # durations - list the slowest test durations
