--- src/promptflow/tests/tracing_test/.coveragerc
@@ -1,19 +1,0 @@
-[run]
-source =
-    */promptflow/tracing/*
-omit =
-    */promptflow/_cli/*
-    */promptflow/_core/*
-    */promptflow/_sdk/*
-    */promptflow/_telemetry/*
-    */promptflow/_utils/*
-    */promptflow/azure/*
-    */promptflow/batch/*
-    */promptflow/contracts/*
-    */promptflow/core/*
-    */promptflow/entities/*
-    */promptflow/executor/*
-    */promptflow/integrations/*
-    */promptflow/operations/*
-    */promptflow/storage/*
-    *__init__.py*
